Module gotham::error

The error module is nascent. At present, it re-exports types from the failure crate and adds an alias for compatible errors. Future directions for Gotham error types are an ongoing discussion. Feel free to chip in.

Structs

Error

The Error type, which can contain any failure.

Type Definitions

CompatError

An alias for failure::Error.compat(), which exists to fulfill the std::error::Error trait.

Result

An implementation of the single-parameter Result pattern, using our pub use failure::Error